ecommerce_api/requirements.txt
asgiref==3.8.1
boto3==1.35.21
botocore==1.35.21
dj-database-url==2.2.0
Django==5.1
django-filter==24.3
django-storages==1.14.4
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
